deployed

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Models\Task;
 
+use Illuminate\Support\Facades\Auth;
+
 class TasksController extends Controller
 {
     /**
@@ -13,16 +15,28 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    // getでmessages/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();         // 追加
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);                                 // 追加
+        // 認証を確認し、ログインしている場合のみタスク一覧を表示
+        if (Auth::check()) {
+            // ログインしているユーザーのIDを取得する
+            $userId = Auth::user()->id;
+        
+            // ログインしているユーザーのタスクのみを取得する
+            $tasks = Task::where('user_id', $userId)->get();
+        
+            return view('tasks.index', compact('tasks'));
+        } else {
+            // ログインしていない場合はログイン画面にリダイレクト
+            // return redirect()->route('login');
+            $tasks = null;
+             // メッセージ一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+            
+            
+        }
     }
 
     /**
@@ -34,12 +48,17 @@ class TasksController extends Controller
     // getでmessages/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
-        $task = new Task;
-
-        // メッセージ作成ビューを表示
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+        if (Auth::check()) {
+            $task = new Task;
+    
+            // メッセージ作成ビューを表示
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+        } else {
+            // ログインしていない場合はログイン画面にリダイレクト
+            return redirect('/');
+        }
     }
 
     /**
@@ -57,6 +76,8 @@ class TasksController extends Controller
             'content' => 'required',
             'status' => 'required|max:10'
         ]);
+        
+        $request->merge(['created_by' => Auth::id()]); // ここを追加
 
          // タスクを作成
         $task = new Task;
@@ -66,6 +87,7 @@ class TasksController extends Controller
         $task->due_date = $request->due_date;
         $task->start_date = $request->start_date;
         $task->status = $request->status;
+        $task->user_id = Auth::id(); // ここを追加
         $task->save();
 
         // タスク一覧へリダイレクトさせる
@@ -85,6 +107,15 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
+        // ログインしているユーザーの情報を取得
+        $user = Auth::user();
+    
+        // ログインしているユーザーとタスクのユーザーが一致するか確認
+        if ($task->user_id !== $user->id) {
+            // 一致しない場合はエラーを返す
+            return redirect('/');
+        }
+    
         // タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
@@ -100,9 +131,19 @@ class TasksController extends Controller
     // getでmessages/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
+        // ログインしていない人をhome画面へリダイレクト
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+        
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-
+    
+        // ログインしているユーザーのIDとタスクのユーザーIDが一致しない場合はhome画面へリダイレクト
+        if ($task->user_id != Auth::id()) {
+            return redirect('/');
+        }
+    
         // タスク編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
@@ -119,6 +160,18 @@ class TasksController extends Controller
     // putまたはpatchでmessages/idにアクセスされた場合の「更新処理」
     public function update(Request $request, $id)
     {
+        // ログインしていない場合はログインページにリダイレクトする
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+        
+        // idの値でタスクを検索して取得
+        $task = Task::findOrFail($id);
+    
+        // タスクのユーザーIDとログイン中のユーザーIDが一致しない場合は'/'にリダイレクトさせる
+        if ($task->user_id != Auth::id()) {
+            return redirect('/');
+        }
         
         // バリデーション
         $request->validate([
@@ -150,12 +203,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値で投稿を検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/')->with('success','Delete Successful');
+        }
     }
 }
